+++ components/usb/iot_usbh_cdc/include/iot_usbh_cdc.h
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2024 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2024-2025 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -8,7 +8,7 @@
 #include <stdint.h>
 #include "esp_err.h"
 #include "usb/usb_host.h"
-
+#include "iot_usbh_cdc_type.h"
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -39,7 +39,7 @@ typedef enum {
  * @attention This callback is called from USB Host context, so the CDC device can't be opened here.
  *
  */
-typedef void (*usbh_cdc_new_dev_cb_t)(usb_device_handle_t usb_dev);
+typedef void (*usbh_cdc_new_dev_cb_t)(usb_device_handle_t usb_dev, void *user_data);
 
 /**
  * @brief CDC driver configuration
@@ -51,12 +51,28 @@ typedef struct {
     int task_coreid;                           /*!< Core of the driver's task, Set it to -1 to not specify the core. */
     bool skip_init_usb_host_driver;            /*!< Skip initialization of USB host driver */
     usbh_cdc_new_dev_cb_t new_dev_cb;          /*!< Callback function when a new device is connected */
+    void *user_data;                           /*!< Pointer to user data that will be passed to the callbacks */
 } usbh_cdc_driver_config_t;
 
 /**
  * @brief Callback structure for CDC device events
+ *
+ * @param[in] cdc_handle Handle to the CDC device that sent the notification
+ * @param[in] user_data User data pointer that was passed during callback registration
  */
 typedef void (*usbh_cdc_event_cb_t)(usbh_cdc_handle_t cdc_handle, void *user_data);
+
+/**
+ * @brief Callback function for CDC device notifications
+ *
+ * This callback is invoked when a CDC device sends a notification event. The notification
+ * data is passed through the notif parameter.
+ *
+ * @param[in] cdc_handle Handle to the CDC device that sent the notification
+ * @param[in] notif Pointer to the notification data structure containing notification details
+ * @param[in] user_data User data pointer that was passed during callback registration
+ */
+typedef void (*usbh_cdc_notif_cb_t)(usbh_cdc_handle_t cdc_handle, iot_cdc_notification_t *notif, void *user_data);
 
 /**
  * @brief Callback structure for CDC device events
@@ -65,7 +81,8 @@ typedef void (*usbh_cdc_event_cb_t)(usbh_cdc_handle_t cdc_handle, void *user_dat
 typedef struct {
     usbh_cdc_event_cb_t connect;            /*!< USB connect callback, set NULL if use */
     usbh_cdc_event_cb_t disconnect;         /*!< USB disconnect callback, set NULL if not use */
-    usbh_cdc_event_cb_t revc_data;          /*!< USB receive data callback, set NULL if not use */
+    usbh_cdc_event_cb_t recv_data;          /*!< USB receive data callback, set NULL if not use */
+    usbh_cdc_notif_cb_t notif_cb;           /*!< USB notification callback, set NULL if not use */
     void *user_data;                        /*!< Pointer to user data that will be passed to the callbacks */
 } usbh_cdc_event_callbacks_t;
 
@@ -161,6 +178,31 @@ esp_err_t usbh_cdc_create(const usbh_cdc_device_config_t *config, usbh_cdc_handl
 esp_err_t usbh_cdc_delete(usbh_cdc_handle_t cdc_handle);
 
 /**
+ * @brief Send a custom control request to the USB CDC device
+ *
+ * This function sends a custom control request to the USB CDC device with the specified parameters.
+ * It can be used for both IN and OUT transfers.
+ *
+ * @param[in] cdc_handle The CDC device handle
+ * @param[in] bmRequestType The request type bitmask
+ * @param[in] bRequest The request ID
+ * @param[in] wValue The request value
+ * @param[in] wIndex The request index
+ * @param[in] wLength The length of data to transfer
+ * @param[in,out] data Pointer to the data buffer:
+ *                     - For OUT transfers: data to be sent to device
+ *                     - For IN transfers: buffer to store received data
+ *
+ * @return
+ *     - ESP_OK: Request sent successfully
+ *     - ESP_ERR_INVALID_ARG: Invalid arguments (NULL handle/data buffer)
+ *     - ESP_ERR_INVALID_STATE: Device is not connected
+ *     - ESP_ERR_TIMEOUT: Control transfer timed out
+ *     - ESP_ERR_INVALID_RESPONSE: Control transfer failed
+ */
+esp_err_t usbh_cdc_send_custom_request(usbh_cdc_handle_t cdc_handle, uint8_t bmRequestType, uint8_t bRequest, uint16_t wValue, uint16_t wIndex, uint16_t wLength, uint8_t *data);
+
+/**
  * @brief Write data to the USB CDC device
  *
  * This function writes data to the specified USB CDC device by pushing the data into the output ring buffer.
@@ -236,6 +278,33 @@ esp_err_t usbh_cdc_flush_tx_buffer(usbh_cdc_handle_t cdc_handle);
  *     - ESP_ERR_INVALID_ARG: Invalid CDC handle provided
  */
 esp_err_t usbh_cdc_get_rx_buffer_size(usbh_cdc_handle_t cdc_handle, size_t *size);
+
+/**
+ * @brief Register an additional callback function for new device detection
+ *
+ * This function registers a callback that will be invoked when a new USB CDC device is connected.
+ * The callback will be automatically unregistered when usbh_cdc_driver_uninstall() is called.
+ *
+ * @param[in] new_dev_cb Callback function to be called when a new device is connected
+ * @param[in] user_data User data pointer that will be passed to the callback function
+ *
+ * @return
+ *     - ESP_OK: Callback registered successfully
+ *     - ESP_ERR_INVALID_ARG: Invalid CDC handle or callback function
+ *     - ESP_ERR_NO_MEM: Failed to allocate memory for callback registration
+ */
+esp_err_t usbh_cdc_register_new_dev_cb(usbh_cdc_new_dev_cb_t new_dev_cb, void *user_data);
+
+/**
+ * @brief Unregister the new device callback which same as the one registered by usbh_cdc_register_new_dev_cb
+ *
+ * @param[in] new_dev_cb The new device callback function to unregister
+ *
+ * @return
+ *     - ESP_OK: Callback unregistered successfully
+ *     - ESP_ERR_INVALID_ARG: Invalid callback function
+ */
+esp_err_t usbh_cdc_unregister_new_dev_cb(usbh_cdc_new_dev_cb_t new_dev_cb);
 
 /**
  * @brief Get the connect state of given interface