+++ examples/usb/host/usb_rndis_4g_module/components/app_wifi/include/app_wifi.h
@@ -1,0 +1,20 @@
+/*
+ * SPDX-FileCopyrightText: 2022-2023 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef _APP_WIFI_H_
+#define _APP_WIFI_H_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void app_wifi_main();
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif